Add files via upload
BACKGROUND: Worldwide Tuberculosis is the major burden to the population. To see the
pattern in annual incidence of tuberculosis statistical model is created.
OBJECTIVES: 1) To study the pattern in annual TB incidence number and annual TB and
HIV incidence number of south Asian countries.2) To study the pattern in annual TB
mortality number and with HIV confounding mortality number of south Asian countries.3)
To predict the annual incidence rate of TB.4) To predict the annual mortality rate of TB.
METIRIALS AND METHODOLOGY: The present study is based on the secondary data
gathered under the program of open data platform and stored in information system. The data
is from 2000 to 2020.South Asian countries are considered in this study. Incidence rate,
Mortality rate, and confounding with HIV is the main variables, are predicted using Time
series analysis. ARIMA model is used for prediction.
CONCLUSION: The annual incidence number of TB, Incidence of TB and HIV, mortality
number of TB cases excluding HIV, mortality number of TB and HIV cases, mortality
number of TB of South Asian countries were predicted for 5 years from 2021 to 2025.The
incidence rate of TB cases in Afghanistan, Bangladesh, Pakistan, and Sri Lanka is increasing.
There is a decrease in the trend in India and Nepal. There is variation in the incidence rate of
Bhutan. The incidence rate and HIV cases in Bangladesh and Pakistan show an increasing
rate, and Afghanistan, Nepal, and Sri Lanka show that there is variation in incidence and HIV
rate. It is rapidly declining in India and Bhutan.